requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas>=2.0.0
flask==3.0.0
plotly==5.17.0
dash==2.16.1
dash-bootstrap-components==1.5.0
gunicorn==21.2.0 